+++ src/emitters/sunsky.cpp
@@ -178,7 +178,7 @@ public:
 		Spectrum value = 
 			sunRadiance * (2 * M_PI * (1-std::cos(theta))) *
 			(bitmap->getWidth() * bitmap->getHeight()) 
-			/ (2 * M_PI * M_PI * nSamples);
+			/ (2.0f * M_PI * M_PI * (Float) nSamples);
 
 		for (size_t i=0; i<nSamples; ++i) {
 			Vector dir = sunFrame.toWorld(
